feat: BugsnagJavascriptSpansPlugin react native implementation
